android/gradle/Homebrew/app/requirements.txt
requests
toga-android~=0.3.0
